+++ bot/models/bump.py
