+++ bot/models/lavalink_result.py
