+++ bot/models/queue_item.py
