+++ bot/server/views/discordoauth.py
@@ -9,7 +9,7 @@ from aiohttp import web
 from aiohttp_session import get_session
 from requests.exceptions import HTTPError, Timeout
 
-from bot.dataclass.oauth import OAuth
+from bot.models.oauth import OAuth
 from bot.utils.constants import DISCORD_API_BASE_URL, USER_AGENT
