+++ bot/server/views/spotifyoauth.py
@@ -10,7 +10,7 @@ from aiohttp import web
 from aiohttp_session import get_session
 from requests.exceptions import HTTPError, Timeout
 
-from bot.dataclass.oauth import OAuth
+from bot.models.oauth import OAuth
 from bot.utils.constants import (
   SPOTIFY_ACCOUNTS_BASE_URL,
   SPOTIFY_API_BASE_URL,
